requirements.txt
pyOpenSSL>=7.5.0
#pyasn1>=0.4.2
python-dateutil>=2.6.1
pytest>=3.0.0
